Add script to wait for the model to appear when using KServe Modelcar
When using OCI containers for model storage in KServe (modelcar), there
is the possibility that the model server starts before the model has
been fully downloaded. When this happens, the model server would
terminate with error because the model path is empty.

This adds a small script to wait for the cluster to fully download the
model container before invoking the model server. The waiting is
triggered when the MODEL_INIT_MODE environment variable is set to the "async".

Signed-off-by: Edgar Hernández <23639005+israel-hdez@users.noreply.github.com>